Refactored the platform package to confirm to new naming standard.
