+++ common/src/main/java/org/aggregatortech/dindora/common/CommonConfigProperty.java
@@ -1,0 +1,7 @@
+package org.aggregatortech.dindora.common;
+
+public enum CommonConfigProperty implements ConfigProperty {
+  SERVICE_DESCRIPTION,
+  SERVICE_PORT;
+
+}
