+++ common/src/main/java/org/aggregatortech/dindora/common/ConfigProperty.java
@@ -1,0 +1,4 @@
+package org.aggregatortech.dindora.common;
+
+public interface ConfigProperty {
+}
