+++ common/src/main/java/org/aggregatortech/dindora/common/Constants.java
@@ -1,0 +1,7 @@
+package org.aggregatortech.dindora.common;
+
+public class Constants {
+
+  public static final String BASE_PACKAGE = "org.aggregatortech.dindora";
+
+}
