+++ common/src/main/java/org/aggregatortech/dindora/exception/ProcessingException.java
@@ -1,0 +1,7 @@
+package org.aggregatortech.dindora.exception;
+
+public class ProcessingException extends BaseException {
+  public ProcessingException(String errorCode) {
+    super(errorCode);
+  }
+}
