--- common/src/main/java/org/aggregatortech/dindora/exception/object/ProcessingException.java
@@ -1,7 +1,0 @@
-package org.aggregatortech.dindora.exception.object;
-
-public class ProcessingException extends BaseException {
-  public ProcessingException(String errorCode) {
-    super(errorCode);
-  }
-}
