--- common/src/main/java/platform/common/CommonConfigProperty.java
@@ -1,7 +1,0 @@
-package platform.common;
-
-public enum CommonConfigProperty implements ConfigProperty {
-  SERVICE_DESCRIPTION,
-  SERVICE_PORT;
-
-}
