--- common/src/main/java/platform/common/ConfigProperty.java
@@ -1,4 +1,0 @@
-package platform.common;
-
-public interface ConfigProperty {
-}
