--- common/src/main/java/platform/common/Constants.java
@@ -1,7 +1,0 @@
-package platform.common;
-
-public class Constants {
-
-  public static final String PLATFORM_PACKAGE = "org.aggregatortech.dindora";
-
-}
