+++ gradle/script/web.gradle
@@ -1,4 +1,4 @@
 apply plugin: 'application'
-mainClassName = "platform.web.service.Server"
+mainClassName = "org.aggregatortech.dindora.web.service.Server"
 
 applicationDefaultJvmArgs = ["-DservicePort=${servicePort}"]
