+++ persistence-file/src/main/java/org.aggregatortech.dindora.persistence/FilePersistenceConfigProperty.java
@@ -1,6 +1,6 @@
 package org.aggregatortech.dindora.persistence;
 
-import platform.common.ConfigProperty;
+import org.aggregatortech.dindora.common.ConfigProperty;
 
 public enum FilePersistenceConfigProperty implements ConfigProperty {
   FILE_PERSISTENCE_LOCATION;
