+++ persistence-file/src/main/java/org.aggregatortech.dindora.persistence/FilePersistenceLocationService.java
@@ -2,8 +2,7 @@ package org.aggregatortech.dindora.persistence;
 
 import org.glassfish.hk2.api.ServiceLocator;
 import org.jvnet.hk2.annotations.Service;
-import platform.common.ServiceLocatorHelper;
-import platform.common.io.system.SystemHelper;
+import org.aggregatortech.dindora.common.io.system.SystemHelper;
 
 import javax.inject.Inject;
