+++ topic/src/integrationTest/groovy/org/aggregatortech/dindora/topic/persistence/TopicFilePersistenceServiceTest.groovy
@@ -1,13 +1,13 @@
 package org.aggregatortech.dindora.topic.persistence
 
 import org.aggregatortech.dindora.common.service.IdGenerationService
-import org.aggregatortech.dindora.exception.object.ProcessingException
+import org.aggregatortech.dindora.common.test.BaseSpecification
+import org.aggregatortech.dindora.exception.ProcessingException
 import org.aggregatortech.dindora.persistence.FilePersistenceLocationService
 import org.aggregatortech.dindora.topic.object.Topic
 import org.glassfish.hk2.api.ServiceLocator
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
-import platform.common.test.BaseSpecification
 
 class TopicFilePersistenceServiceTest extends BaseSpecification {
     @Rule
